requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
yt-dlp
requests
python-dotenv
isodate
openpyxl
